article then

+++ js/ArticleService.js
@@ -5,7 +5,7 @@ const baseUrl = 'https://panda-market-api-crud.vercel.app/articles';
 
 
 // 아티클 리스트 조회
-async function getArticleList(page, pageSize, keyword) {
+function getArticleList(page, pageSize, keyword) {
     let url = baseUrl;
 
     url += '?';
@@ -17,54 +17,80 @@ async function getArticleList(page, pageSize, keyword) {
     // url.searchParames.append('pageSize', pageSize);
     // url.searchParames.append('keyword', keyword);
 
-    const response = await fetch(url);
+    const response = fetch(url);
 
-    if (!response.ok) {
-        switch (response.status) {
-            case 404:
-                throw new Error("존재하지 않는 데이터입니다.");
-            case 403:
-                throw new Error("데이터를 조회할 권한이 없습니다.");
-            default:
-                throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+    const data = response.then((response) => {
+        if (!response.ok) {
+            switch (response.status) {
+                case 404:
+                    throw new Error("존재하지 않는 데이터입니다.");
+                case 403:
+                    throw new Error("데이터를 조회할 권한이 없습니다.");
+                default:
+                    throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+            }
         }
-    }
 
-    const data = await response.json();
+        return response.json();
 
-    const arrData = data.list;
+    }).then((data) => {
 
-    return arrData;
+        const arrData = data.list;
+
+        return arrData;
+    })
+
+    return data;
 }
 
 
 
 // 아티클 리스트 상세
-async function getArticle(id) {
+function getArticle(id) {
+    
+    // 필수값 확인 후 없으면 리젝시키기
+    if (id === undefined || id === null) {
+        return Promise.reject(new Error("조회할 아티클 id가 없습니다."));
+    }
+
     let url = baseUrl + '/' + id;
     
-    const response = await fetch(url);
+    const response = fetch(url);
 
-    if (!response.ok) {
-        switch (response.status) {
-            case 404:
-                throw new Error("존재하지 않는 데이터입니다.");
-            case 403:
-                throw new Error("데이터를 조회할 권한이 없습니다.");
-            default:
-                throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+    const data = response.then((response) => {
+
+        if (!response.ok) {
+            switch (response.status) {
+                case 404:
+                    throw new Error("존재하지 않는 데이터입니다.");
+                case 403:
+                    throw new Error("데이터를 조회할 권한이 없습니다.");
+                default:
+                    throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+            }
         }
-    }
-    
-    const data = await response.json();
-    
+        
+        return response.json();
+    })
+       
     return data;
 }
 
 
 
 // 아티클 신규 생성
-async function createArticle(title, content, image) {
+function createArticle(title, content, image) {
+
+    if (title === undefined || title === null) {
+        return Promise.reject(new Error("타이틀이 없어 아티클을 생성할 수 없습니다."));
+    }
+    if (content === undefined || content === null) {
+        return Promise.reject(new Error("콘텐트가 없어 아티클을 생성할 수 없습니다."));
+    }
+    if (image === undefined || image === null) {
+        return Promise.reject(new Error("이미지가 없어 아티클을 생성할 수 없습니다."));
+    }
+
     let url = baseUrl;
 
     const bodyContent = {
@@ -73,7 +99,7 @@ async function createArticle(title, content, image) {
         image
     }
 
-    const response = await fetch(url, {
+    const response = fetch(url, {
         method: 'POST',
         body: JSON.stringify(bodyContent),
         headers: {
@@ -81,19 +107,28 @@ async function createArticle(title, content, image) {
         },
     })
 
-    if (!response.ok) {
-        throw new Error("데이터를 생성하는데 실패했습니다..");
-    }
+    const data = response.then((response) => {
 
-
-    const data = await response.json();
-
+        if (!response.ok) {
+            throw new Error("데이터를 생성하는데 실패했습니다..");
+        }
+        
+        return response.json();
+    })
+       
     return data;
+    
 }
 
 
 // 아티클 내용 수정
-async function patchArticle(id, title = null, content = null, image = null) {
+function patchArticle(id, title = null, content = null, image = null) {
+    
+    // 필수값 확인 후 없으면 리젝시키기
+    if (id === undefined || id === null) {
+        return Promise.reject(new Error("수정할 아티클 id가 없습니다."));
+    }
+
     let url = baseUrl + '/' + id;
 
     const bodyContent = {}
@@ -108,7 +143,7 @@ async function patchArticle(id, title = null, content = null, image = null) {
         bodyContent.image = image;
     }
 
-    const response = await fetch(url, {
+    const response = fetch(url, {
         method: 'PATCH',
         body: JSON.stringify(bodyContent),
         headers: {
@@ -116,39 +151,56 @@ async function patchArticle(id, title = null, content = null, image = null) {
         },
     })
 
-    if (!response.ok) {
-        switch (response.status) {
-            case 404:
-                throw new Error("존재하지 않는 데이터입니다.");
-            default:
-                throw new Error("요청하신 데이터를 찾을 수 없습니다.");
-        }
-    }
 
-    const data = await response.json();
+    const data = response.then((response) => {
+
+        if (!response.ok) {
+            switch (response.status) {
+                case 404:
+                    throw new Error("존재하지 않는 데이터입니다.");
+                case 403:
+                    throw new Error("데이터를 조회할 권한이 없습니다.");
+                default:
+                    throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+            }
+        }
+        
+        return response.json();
+    })
 
     return data;
 }
 
-async function deleteArticle(id) {
+function deleteArticle(id) {
+
+    // 필수값 확인 후 없으면 리젝시키기
+    if (id === undefined || id === null) {
+        return Promise.reject(new Error("삭제할 아티클 id가 없습니다."));
+    }
+
     let url = baseUrl + '/' + id;
 
-    const response = await fetch(url, {
+    const response = fetch(url, {
         method: 'DELETE',
     })
 
-    if (!response.ok) {
-        switch (response.status) {
-            case 404:
-                throw new Error("존재하지 않는 데이터입니다.");
-            case 403:
-                throw new Error("데이터를 조회할 권한이 없습니다.");
-            default:
-                throw new Error("요청하신 데이터를 찾을 수 없습니다.");
-        }
-    }
+    const data = response.then((response) => {
 
-    const data = await response.json();
+        if (!response.ok) {
+            switch (response.status) {
+                case 400:
+                    throw new Error("400");
+                case 404:
+                    throw new Error("존재하지 않는 데이터입니다.");
+                case 403:
+                    throw new Error("데이터를 조회할 권한이 없습니다.");
+                default:
+                    throw new Error("요청하신 데이터를 찾을 수 없습니다.");
+            }
+        }
+        
+        return response.json();
+    })
 
     return data;
 }
